Ajuste no layout da tela de vinculo. Ajuste para estar excluindo as pastas anterior a um dia no vnculo e trocando o tipo de nomenclatura da pagina. Também estará mantendo as 10 ultimas ativa do manga a ser vinculado. Ajuste para o recarregamento das imagnes e aparição de um botão para recarregar quando não conseguir em 3 tentativas.

+++ app/src/main/java/br/com/fenix/bilingualmangareader/service/parses/RarParse.kt
@@ -135,15 +135,18 @@ class RarParse : Parse {
         }
     }
 
-    override fun destroy() {
-        if (mCacheDir != null) {
-            for (f in mCacheDir?.listFiles()!!)
-                f.delete()
-
-            mCacheDir!!.delete()
+    override fun destroy(isClearCache: Boolean) {
+        if (isClearCache) {
+            if (mCacheDir != null) {
+                mCacheDir?.listFiles()?.let {
+                    for (f in it)
+                        f.delete()
+                }
+                mCacheDir?.delete()
+            }
         }
         mHeaders.clear()
-        mArchive!!.close()
+        mArchive?.close()
         mArchive = null
     }
 
@@ -153,13 +156,14 @@ class RarParse : Parse {
 
     fun setCacheDirectory(cacheDirectory: File?) {
         mCacheDir = cacheDirectory
-        if (!mCacheDir!!.exists())
-            mCacheDir!!.mkdirs()
+        mCacheDir?.let {
+            if (!it.exists())
+                it.mkdirs()
 
-        if (mCacheDir!!.listFiles() != null) {
-            for (f in mCacheDir?.listFiles()!!)
-                f.delete()
-
+            if (it.listFiles() != null) {
+                for (f in it.listFiles()!!)
+                    f.delete()
+            }
         }
     }
 }
